Project updated
Implemented bank account adding feature;
Implemented overall financial statistcis sum search feature;
Changed request mapping for several requests;

+++ src/main/java/org/paperplane/organizationsregister/data/FinancialStatisticsRepository.java
@@ -16,7 +16,7 @@ public interface FinancialStatisticsRepository extends JpaRepository<FinancialSt
     FinancialStatisticsByQuarter findById(int id);
     List<FinancialStatisticsByQuarter> findAllByOrganizationAndYear(Organization organization, int year);
     List<FinancialStatisticsByQuarter> findAllDebit();
-    FinancialStatisticsByQuarter findByOrganizationAndYearAndQuarter(Organization organization, int year, byte quarter);
+    FinancialStatisticsByQuarter findByOrganizationAndYearAndQuarter(Organization organization, int year, int quarter);
     BigDecimal sumFinancesOfOrganizationWithinYear(Organization organization, int year);
     List<FinancialStatisticsByQuarter> findAllByOrganization(Organization organization);
     boolean existsById(Integer id);
